+++ src/classes/equip/Potion.cpp
@@ -2,8 +2,8 @@
 
 
 Potion::Potion(Level &level, string &fileName, string &name, string &type, string &subType, float x, float y, float w,
-               float h, int state, const vector<pair<string, int>> &changesList) : Item(level, fileName, name, type,
-                                                                                        subType,
-                                                                                        x, y, w, h, state) {
+               float h, int state, const vector<pair<string, float>> &changesList) : Item(level, fileName, name, type,
+                                                                                          subType,
+                                                                                          x, y, w, h, state) {
     this->changesList = changesList;
 }
